LayoutTests/fast/js/sputnik/Conformance/08_Types/8.7_The_Reference_Type/S8.7.2_A1_T2-expected.txt

S8.7.2_A1_T2

PASS ReferenceError: Left side of assignment is not a reference.

TEST COMPLETE
